# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface AppleSubscriptionPrice {
  productId: string;
  /** Storefront country code in the same form as the request. */
  territory: string;
  currencyCode: string;
  localizedPrice: string;
}

export interface HealthStatus {
  status: string;
}

export type GetAppleSubscriptionPriceParams = {
productId: string;
/**
 * StoreKit storefront country code (ISO alpha-3); alpha-2 is accepted for older clients.
 * @pattern ^[A-Z]{2,3}$
 */
countryCode: string;
};