# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
paths:
  /billing/apple-price:
    get:
      operationId: getAppleSubscriptionPrice
      summary: Get Apple's current standard subscription price for an App Store storefront
      description: Authenticated. The product must belong to this app's Apple subscriptions. Returns 503 rather than guessing if Apple pricing cannot be verified.
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            type: string
        - name: countryCode
          in: query
          required: true
          description: StoreKit storefront country code (ISO alpha-3); alpha-2 is accepted for older clients.
          schema:
            type: string
            pattern: "^[A-Z]{2,3}$"
      responses:
        "200":
          description: Current standard new-subscriber price from Apple
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppleSubscriptionPrice"
        "400":
          description: Invalid query
        "503":
          description: Price cannot be verified
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  schemas:
    AppleSubscriptionPrice:
      type: object
      properties:
        productId:
          type: string
        territory:
          type: string
          description: Storefront country code in the same form as the request.
        currencyCode:
          type: string
        localizedPrice:
          type: string
      required: [productId, territory, currencyCode, localizedPrice]
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status