# lib/api-zod/src/generated/types/appleSubscriptionPrice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppleSubscriptionPrice {
  productId: string;
  /** Storefront country code in the same form as the request. */
  territory: string;
  currencyCode: string;
  localizedPrice: string;
}